pyproject.toml
[project]
name = "agentspace-agent-registrar"
version = "0.1.0"
description = "Manage agents in Agentspace Agent Gallery and Agent Engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.40.3",
    "google-cloud-aiplatform>=1.103.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
]

[project.scripts]
agentspace-registrar = "agentspace_registrar.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.11",
]
